+++ Changelog.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+- Do not set resource titles
+- Use `summary` instead of `operationId` for action titles
+- Set a relation value based on `operationId` for actions
+
 # 0.1.0 - 2015-09-22
 
 - Initial release.